packages/templates/src/research-agent/templates/langchain-python/requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0

# Web tools
requests>=2.31.0
beautifulsoup4>=4.12.0

# Environment variables
python-dotenv>=1.0.0

# Optional: Search API integrations (uncomment as needed)
# google-search-results>=2.4.2  # For SerpAPI
# tavily-python>=0.3.0          # For Tavily Search
